requirements.txt
# Core ML/DL
torch>=2.0.0,<3.0.0
transformers>=4.36.0,<5.0.0
datasets>=2.14.0,<3.0.0
accelerate>=0.24.0,<1.0.0

# Fine-tuning & Quantization
peft>=0.6.0,<1.0.0
bitsandbytes>=0.41.0,<1.0.0
trl>=0.7.0,<1.0.0

# MLOps & Tracking
mlflow>=2.8.0,<3.0.0
dagshub>=0.3.0,<1.0.0

# Utilities
pyyaml>=6.0,<7.0
python-dotenv>=1.0.0,<2.0.0
huggingface_hub>=0.19.0,<1.0.0
sentencepiece>=0.1.99,<1.0.0
protobuf>=3.20.0,<5.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Evaluation
evaluate>=0.4.0,<1.0.0
rouge_score>=0.1.2,<1.0.0

# System utilities
psutil>=5.9.0,<6.0.0

# Testing
pytest>=7.0.0,<9.0.0
pytest-cov>=4.0.0,<6.0.0
